+++ packages/imperative/src/config/__tests__/Config.api.unit.test.ts
@@ -216,16 +216,16 @@ describe("Config API tests", () => {
                 expect(profile).toBeNull();
             });
         });
-        describe("expandPath", () => {
+        describe("getProfilePathFromName", () => {
             it("should expand a short proeprty path", async () => {
                 const config = await Config.load(MY_APP);
                 const profilePath = "lpar1.zosmf";
-                expect(config.api.profiles.expandPath(profilePath)).toEqual("profiles.lpar1.profiles.zosmf");
+                expect(config.api.profiles.getProfilePathFromName(profilePath)).toEqual("profiles.lpar1.profiles.zosmf");
             });
             it("should expand a path with the keyword profiles", async () => {
                 const config = await Config.load(MY_APP);
                 const profilePath = "profiles.zosmf";
-                expect(config.api.profiles.expandPath(profilePath)).toEqual("profiles.profiles.profiles.zosmf");
+                expect(config.api.profiles.getProfilePathFromName(profilePath)).toEqual("profiles.profiles.profiles.zosmf");
             });
         });
         describe("getProfileNameFromPath", () => {