+++ packages/cli/CHANGELOG.md
@@ -2,10 +2,16 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Changed references in command output from 'Team Configuration' to 'Zowe client configuration' [#2019](https://github.com/zowe/zowe-cli/issues/2019).
+
 ## `8.0.0-next.202403132009`
+
 - Enhancement: Prompt for user/password on SSH commands when a token is stored in the config. [#2081](https://github.com/zowe/zowe-cli/pull/2081)
 
 ## `8.0.0-next.202403131702`
+
 - BugFix: Removing stack trace for zosjobs errors. [#2078](https://github.com/zowe/zowe-cli/pull/2078)
 
 ## `8.0.0-next.202403122137`